Removed interfaces with a single implementation
