+++ docker-adapter/src/test/java/com/artipie/docker/asto/AstoManifestsTest.java
@@ -9,11 +9,11 @@ import com.artipie.asto.Content;
 import com.artipie.asto.Storage;
 import com.artipie.docker.Blob;
 import com.artipie.docker.ExampleStorage;
+import com.artipie.docker.ManifestReference;
 import com.artipie.docker.RepoName;
 import com.artipie.docker.Tags;
 import com.artipie.docker.error.InvalidManifestException;
 import com.artipie.docker.manifest.Manifest;
-import com.artipie.docker.ManifestReference;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.hamcrest.core.IsEqual;
